refactor: ItemList를 삭제, CalculatorItemQueue를 링크드 리스트로 활용
