+++ Calculator/Calculator/Model/CalculateItem.swift
@@ -8,3 +8,5 @@
 protocol CalculateItem {
     
 }
+
+extension String: CalculateItem { }
