+++ Calculator/Calculator/Model/CalculatorItemQueue.swift
@@ -5,15 +5,43 @@
 //  Created by 이정민 on 2022/09/19.
 //
 
-struct CalculatorItemQueue: CalculateItem {
-    var queue: ItemList = ItemList()
+struct CalculatorItemQueue<Item: CalculateItem> {
+    private(set) var head: Node<Item>?
+    private(set) var tail: Node<Item>?
+    private(set) var count: Int = 0
     
-    mutating func push(_ data: String) {
-        queue.enqueue(data)
+    mutating func enqueue(_ data: Item) {
+        let newNode = Node(data: data)
+
+        if isEmpty() {
+            self.head = newNode
+        } else {
+            self.tail?.next = newNode
+        }
+        
+        self.tail = newNode
+        count += 1
     }
     
-    @discardableResult
-    mutating func pop() -> String? {
-        queue.dequeue()
+    mutating func dequeue() -> Item? {
+        guard let head = self.head else {
+            return nil
+        }
+        
+        let data = head.data
+        self.head = self.head?.next
+        count -= 1
+        
+        return data
+    }
+    
+    mutating func clearQueue() {
+        self.head = nil
+        self.tail = nil
+        count = 0
+    }
+    
+    func isEmpty() -> Bool {
+        self.head == nil ? true : false
     }
 }
